requirements.txt
# requirements.txt - Install these with pip install -r requirements.txt
# On Raspberry Pi, ensure Python 3.9+ and pip are installed.
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
mariadb==1.1.10  # For MariaDB connector
pydantic==2.0.3
authlib==1.3.2  # For Google OAuth
python-dotenv==1.0.1
jinja2==3.1.4
passlib[bcrypt]==1.7.4  # For password hashing (admin login)
python-jose[cryptography]==3.3.0  # For JWT
httpx==0.27.2